requirements.txt
h5py==2.6.0
numpy==1.12.0
protobuf==3.1.0.post1
PyYAML==3.12
scipy==0.18.1
six==1.10.0
tensorflow-gpu==1.0.0rc0
click>=6.7
